Move reader_parser to arrow-cast (#3022)
